fix a deprecated int-char comparison in test/perf/shootout

+++ test/perf/shootout/revcomp.jl
@@ -47,7 +47,7 @@ function revcomp(infile="revcomp-input.txt")
         if isempty(line)
 #            print_buff(buff)
             return
-        elseif line[1] == '>'
+        elseif line[1] == UInt8('>')
 #            print_buff(buff)
             buff = UInt8[]
 #            write(line)
